use 'oc create' to create route as a workaround for 'oc expose' bug
There is currently a bug with Endpoint Ports when using 'oc expose' :

https://bugzilla.redhat.com/show_bug.cgi?id=1431781

The Endpoint Port for the route is set to the svc port instead of pods.
The Pull Request is not yet merged : https://github.com/openshift/origin/pull/13609

Use 'oc create' as a workaround as it does set the Endpoint port properly to
'<all endpoint ports>'.